move tab components back into individual parts
