+++ uikit/Tabs/stories.tsx
@@ -1,7 +1,7 @@
 import { storiesOf } from '@storybook/react';
 import React from 'react';
 import Tabs, { Tab } from '.';
-import Typography from '../../Typography';
+import Typography from '../Typography';
 import { action } from '@storybook/addon-actions';
 
 function TabContainer(props) {
